implémentation du controler a la place du main, ajout des différentes méthodes et des interfaces GUI et Controler. Correction du Model pour accepter des doubles

+++ calculatrice/src/main/java/calculatricenpi/Model/CalculatorModelInterface.java
@@ -8,7 +8,7 @@ public interface CalculatorModelInterface {
     public void divide();
     public void opposite();
     public void push();
-    public int pop();
+    public double pop();
     public void drop();
     public void swap();
